TableBuilder spec: Add extra links from `ReferentialDecorator`
A couple more links should appear in the gear menu thanks to the
inclusion of the `ReferentialDecorator#action_links`'s links in the
menu.

The test still fails, though, because the destroy link that comes out of
the decorator isn't formatted correctly (it looks like all the other
links instead of having a special icon).

Refs #3479