TableBuilder: Make destroy link in gear menu render correctly
Previously we weren't correctly outputting the proper HTML to get the
:delete link rendering correctly with the icon and text. It should show:

    [TRASH CAN ICON] Supprimer

but instead was showing the href.

To get this working, I added a `@content` attribute to `Link`. This
allows us to optionally specify non-text content to display in the link.

We then have delete-link-specific handling when rendering the gear
menu's HTML. In that particular case, the `<li>` needs to have a
'delete-action' class in order for the styling to work properly.

Created a new `destroy_link` helper method to allow us to reuse the
destroy link content in the header bar.

TODO: Collect Link#content and Link#name into the same property.
TODO: Rename `#destroy_link` to `#destroy_link_content`

Refs #3479