TableBuilderHelper#table_builder_2: Clean up arguments & add docs
Remove unnecessary arguments and document the ones we are using.

`current_referential` has been removed. Ultimately decided that it was
simpler to just rely on the existing method that's magically "imported"
instead of passing it directly. If you vote to pass it in as an arg,
though, I'm all for it.

Refs #3479