referential_lines/show.html.slim: Remove TODO
I had added this as an idea for how to render these links both here, in
the page header, and in the table gear menu of the preceding page. We've
basically done that now with `TableBuilderHelper`, although in a
slightly different way than I originally imagined, with decorators (like
`ReferentialDecorator#action_links`).

Refs #3479