MultipleSelectionToolbox: Clean up code; Add to workbenches/show.html
Keep the helper code globally the same.

Changes:
* Remove `if action == :edit` since we never execute that branch. In
  fact, the only place we ever use the multiple selection toolbox in
  conjunction with the table builder is on Workbenches#show, and that
  page only uses a :delete action.
* Rename `dPath` to be clearer
* Add some whitespace to improve readability
* Rename `actitem` to `action_link` for better readability
* Connect the links and label with `#+` instead of `#concat`

Add the toolbox to "workbenches/show.html.slim" to get the page looking
like it was before I changed the table builder.

Needed to wrap the table and toolbox in a `<div class="select_table">`
in order for the JavaScript part of it to work (otherwise selection
wouldn't activate the link, and likely other stuff would be messed up
too).

Now that the table builder and the selection toolbox are separate, it
makes sense to put the "select_table" `<div>` in the template.

Refs #3479