Add new #action_links decorators for a few priority models
These models are being prioritised for changing their views to use the
new TableBuilderHelper.

We're giving them all action links, copied out of the following files:

* time_tables/show.html.slim
* calendars/show.html.slim
* lines/show.html.slim
* routing_constraint_zones/show.html.slim
* routes/show.html.slim

Created a new namespace in decorators/ for Chouette model decorators for
consistency with the model/ directory. We'll want to move
`CompanyDecorator` in there too because it lives in the `Chouette::`
namespace.

Refs #3479