Lines#show: Render header links from #action_links
Get the header links from the decorator so we can reuse them instead of
defining them directly in the template.

Add an option to the `destroy_link_content` Rails helper method to allow
us to specify alternate translation keys, giving us different text in
the button label. Added this because the template in question uses a
different label than the others (usually it's `actions.destroy`).

Refs #3479