referentials_permissions_spec.rb: Fix failure from `#destroy_link_content`
Now, destroy ("Supprimer") buttons in the header get rendered without a
`<span>` tag surrounding the text inside the button. This is because I
consolidated the HTML structure of the delete buttons inside table gear
menus and the header.

The ones in the gear menus had no span, and these in the header do.

In both places, we're now using `LinksHelper#destroy_link_content` to
generate the contents of the button, making the contents uniform in both
contexts.

This means, however, that the markup is not the same for header buttons,
which now have their inner `<span>` tag removed. Update the test to
reflect this.

Also, the styling breaks when I updated the markup, but Jean-Paul
offered to fix that.

Refs #3479