Split factories from chouette_2_factories.rb
Split out all the factories defined in this file into their own separate
files. This helps to make our factories more consistent, so we don't
have a bunch of newer factories in their own files and another bunch of
factories from the early days (presumably) of the application, which are
defined in a single file.

This came about because I was looking for a Referential factory, and was
surprised to not see one defined. Then Xin showed me this grouped
factory file, and I saw that it actually does exist. With this change,
the factories will better match expectations.