TableBuilderHelper: Try not to make cog menu dependent on `links` (WIP)
We do a check for `links.any?` before rendering the cog menu, meaning
that if the `links` argument isn't passed in but there are links in the
model decorator (`#action_links`), those links won't be rendered because
no cog menu will be rendered.

The cog menu should be rendered regardless of whether `links` is passed
in as long as there are links to render.

The problem is that we get certain links from the object itself. I
started work on extracting that to allow us to render the cog menu in
the appropriate cases. This works when the logic stays in `build_links`.
However, the `#thead` method also needs to know how many links there are
to figure out whether it should put an extra `<th>` at the end of the
header column. The trouble is, we don't know at that stage how to get
the links from the item, since we're not looping over items at that
point we only have the collection.

Need to make this work for the header column as well.

Leaving this alone for now as a big WIP commit because I recall that for
now, all `table_builder_2` calls are going to have a `links: [:show]`
argument (at least :show). So this seemed less critical at the moment
than converting tables to the new table builder.

Refs #3479