TableBuilderHelper::URL: Inject `current_referential` dependency
Turns out the only reason why this code worked before was because the
only place I was using `table_builder_2` was for a list of
`Referential`s, and there's an explicit check for
`item.is_a?(Referential)` that avoids calling `current_referential`.

Otherwise, when that `unless` condition passes, we get a failure because
`current_referential` can't be found. It can't be found because helper
methods are not accessible in this scope (duh).

In order to make the referential accessible to the method, require one
as an argument. Now we explicitly pass the `current_referential` from
places where `#polymorphic_url_parts` is called.

Refs #3479