Lines#index: Use new `table_builder_2`
Convert the table of Lines to use the new version of the table builder.
Update the view with the new calling format of the helper.

The controller needs to decorate the `@lines` collection in order to
make `#action_links` available to the table builder.

Needed to update the "_filters.html.slim" template to call
`#human_attribute_name` on the class instead of the collection, because
we can't do that on a decorated collection.

Refs #3479