TableBuilderHelper: Fix NoMethodError when `current_referential` undef
Certain controllers (`Calendar` and `Referential`) don't define a
`#current_referential` method. In those cases, we'd get a
`NoMethodError` in this function when calling `#current_referential`.

To avoid the error, give `current_referential` a default value. Using a
default of `nil` works here because the current referential is needed in
other places in order to build links. Since the links in the special
models don't need a referential in order to be built, we don't care that
the value is `nil`.

Refs #3479