TableBuilderHelper: Fix specs which mock `current_referential`
These two specs were failing because of the line:

    current_referential ||= nil

which would set `current_referential` to `nil` even after updating the
object being stubbed to `helper`.

Change the specs to stub the method on the correct object now that
`TableBuilderHelper::URL` doesn't depend on `current_referential`.

Update the faulty line in question to not clobber `current_referential`,
but still give us the `nil` value we were looking for when the helper
method isn't defined.

Refs #3479