ReferentialLines#show: Use new table builder helper
Convert the table of `Route`s on this page to use the new table builder
helper. This gives us the links from the `RouteDecorator` in the gear
menu next to each row in the table.

Ended up removing the `route_sp` field from the `RouteDecorator` context
because it seems like we can just grab that collection directly from the
`Route` object. At first I thought about renaming it because after
coming back to this code I thought, 'what does sp mean?'. Then after
some investigation, it appears that we don't need to pass it in
explicitly. Now, no need to bother with that from the ReferentialLines
controller.

The RoutesController does create a separate `@route_sp` variable, but
this only changes the sort order of the stop points, not the contents of
the collection. Thus it seems safe to get them directly from the Route
in the decorator.

Refs #3479