TimeTables#index: Use new table builder helper
Additionally convert the `#human_attribute_name` calls to call it on the
class because this is more correct and because that method doesn't exist
on the decorated collection of `@time_tables`.

TODO: There's a bug here where the #show and #edit links don't link
correctly, they go to `/referentials/:id` instead of
`/referentials/:id/time_tables/:id`.

Refs #3479