TableBuilderHelper::URL: Fix polymorphic URL for TimeTable
Had a problem on the TimeTables#index page where the #show and #edit
links in the gear menu would link to `/referentials/:id` instead of
`/referentials/:id/time_tables/:id`.

This turned out to be caused by a syntax bug in the condition that adds
the `Chouette::TimeTable` object to the polymorphic URL array. Now the
proper URL is rendered.

Refs #3479