TableBuilderHelper: Move `referential` variable to method
At Robert's recommendation, use `try` instead of `respond_to?` because
it's shorter.

Also at Robert's recommendation, move the `referential` variable to a
method to memoize it. This allows us to use it in other places without
re-getting it and without having to do the `if` check again.

He had suggested:

    def referential
      return @__referential__ if instance_variable_defined?(:@__referential__)
      @__referential__ = try(:current_referential)
    end

but I think the `||=` version works the same way and is shorter. Still
need to find out from him what our style guide rules for __variables__
are and why I should be using one here.

Refs #3479