Remove ReferentialLines#index
Delete this route and associated template files. Wasn't sure if using
`exclude:` in the routes file was the best way to do this, as I'm not
sure if there are other routes we want to exclude from here also.

I visited the page because I'm migrating templates to the new table
builder helper and this one uses table builder. Upon visiting the page,
it looked pretty wonky, and after consulting with Jean-Paul and Luc, it
turns out this page isn't used any more because the table of
`Chouette::Line`s rendered there is now instead displayed on the
Referentials#show page, obviating the need for this page.

Refs #3479