Referentials#show: Use new table builder helper
Needed to add a new decorator to provide links to this table's gear
menu. The links correspond to those in the header on the
"referential_lines/show.html.slim" page.

Call `#human_attribute_name` on the model directly instead of the
collection because the decorated collection doesn't expose that method.

Refs #3479