ReferentialLines#show: Render links from ReferentialLineDecorator
Leverage our newly created `ReferentialLineDecorator` to render the
header links, allowing us to use them both here and in the table of
lines on the Referentials#show page.

Used the extended decoration syntax to allow us to specify the
`ReferentialLineDecorator`, otherwise Draper will infer `LineDecorator`,
which is not what we want in this case.

Refs #3479