Routes#show: Use new table builder helper
Update the table builder helper for the table of `Chouette::StopPoints`
on this page.

Created a new `StopPointDecorator` which unfortunately has the same code
as `StopAreaDecorator`, the only difference being that instead of using
`object`, it uses `object.stop_area`. Need to work out how to eliminate
that code duplication.

Refs #3479