routes/show.html.erb_spec.rb: Update for new table builder helper
Need to decorate the stop points collection in order to be able to call
`#action_links` on the objects therein.

Added a stub for `pundit_user` to get around
`TableBuilderHelper#table_builder_2` having a dependency on
`ApplicationController#pundit_user`. Don't like this stub at all. Should
be doing this in a better way.

Refs #3479