Add `WorkbenchImport` model
A new model that represents a batched import. The zip archive attached
to this record will contain one or more datasets. A `Referential` will
be created for each dataset in the zip, and that sub-dataset will be
scheduled for a Netex import (only handling Netex imports at the
moment).

The model is currently empty but I'm committing it now so it can be used
by others on the team.

Refs #3507