VehicleJourneyAtStop: Validate day offsets are not greater than 1 day
We want to ensure day offsets can only be used to make at-stops for the
current day and the next day. Any more days are declared to be invalid.

For now, with bus and rail modes of transportation, we don't plan on
having journeys that extend beyond a day after the initial departure.

Also prevent setting negative day offsets while we're at it because it
goes in the same validation block and we should allow have those.

Refs #3597