VehicleJourneyAtStop: Add error message to max day offset error
Add a custom error message for when the day offset exceeds our maximum.
Unfortunately, this uses Rails 5 lambda syntax, so it doesn't work. Will
need to extract it to a method, but figured I'd just commit this anyway
for posterity.

The French error message text was taken from the Redmine issue, authored
by Luc. I translated it into English.

Refs #3597