VehicleJourneyAtStop: Add validation messages to day offset validator
Since we can't use the lambda syntax when passing a message to a
validator, write a custom validation method that allows us to be more
flexible when setting a message. The correct message with the correct
data is now used for the offset error.

This eliminates our validation of `numberiality`, making our two
validation specs no longer valid.

Refs #3597