VehicleJourneyAtStop: Remove "nested" #error_message method
Robert adeptly pointed out that my `#error_message` method is not a
nested method, like I had thought, à la Python. Instead, it actually
defines the `#error_message` method on the class at runtime
(https://stackoverflow.com/questions/4864191/is-it-possible-to-have-methods-inside-methods/4865161#4865161),
so `#error_message` becomes a normal instance method after
`#day_offset_must_be_within_range` is called. I did not know this, and
that's certainly not what I was going for.

Convert the method to a lambda to put it in the local scope.

Refs #3597