VehicleJourneyAtStop: Handle nil value when checking offset within range
Old `VehicleJourneyAtStop`s were created before I added a database
default of 0 to those records. Thus old data will not have the proper 0
value to check against, and our `<` `>` comparisons trigger an error due
to a comparison with `nil`.

To get around the error, set a default value of 0 (the current day) when
the offset isn't set.

I think that ideally this should be solved at the database level with a
data migration, but hopefully this is okay for now. Whenever we do
decide to do a data migration, however, this code needs to be deleted.

Refs #3597