VehicleJourney: Remove 'departure time before arrival time' validation
Remove the `#vjas_departure_time_must_be_before_next_stop_arrival_time`
validation. This prevented vehicle journeys from being saved if they
contained at-stops with a day offset > 0.

For example:

    Arrival | Departure
    23:30   | 23:30
    00:05   | 00:05

should work because 00:05 is on the next day, but it doesn't because the
validation sees it as !(23:30 < 00:05).

Luc says we should remove this validator in order for the day offsets to
work. He instructed me to comment the code instead of removing it
outright, because we're not sure yet what we want to do with this, and
he needs to look into the use-case for the validation. It could be that
we completely remove the validation method, or maybe we modify it
somewhat to play nicely with day offsets.

Skip the tests that check this validation method because we've commented
out the validation. Add a test to check that saving a vehicle journey
with a positive at-stop day offset works (it failed before commenting
out the validation). Unfortunately it seemed too complicated to use
`build_stubbed` in my new test, so I gave up and just used `create`.
Sorry test speedsters. Maybe you can refactor something more
enlightened.

Refs #3597