VehicleJourneyAtStop spec: Remove unnecessary initial values
I had created an at-stop with nil `arrival_day_offset` and
`departure_day_offset` values because that's the kind of object this
test is testing, but when I reviewed the code, it occurred to me that
since the method takes the offset as an argument, it isn't necessary to
provide these attribute values to the factory.

Not sure whether it's clearer this way or the way it was before, but I
figured since they're not necessary it may be better to remove them.

Refs #3597