NetexImport: Validate presence of referential_id & workbench_id
It was possible to create invalid NetexImports via the
NetexImportsController API. These imports had no associated
`workbench_id`. Because of this, the Java import application returned
an error when trying to process the import. This is the response from
the Java server, as printed in the Rails server log:

    Call iev get localhost:8080/boiv_iev/referentials/importer/new?id=4
    IEV server error : can't dup NilClass

Luc noticed the missing attribute problem in our sample data. In order
to ensure the data is getting created correctly, add these validations.

Refs #4155