Import: Add #symbols_with_indifferent_access
Didn't like the `flat_map` written inline in the `#failing_statuses`
method because it wasn't obvious what it was doing.

Got that from this answer by mr.musicman on Stack Overflow:
https://stackoverflow.com/questions/21816139/is-there-something-like-with-indifferent-access-for-arrays-usable-for-include/42723778#42723778

(thank you) that describes how to make an array that responds to
`#include?` for both symbols and string versions of those symbols.

Extract that logic into its own method. Leaving it in `Import` for now
because I only need it here and I didn't want to bother putting it in
another module.

Refs #3509, #3511