Import#child_change: Fix bug when parent status already set
Don't update the parent status if it's already been set to a finished
status. This protects against situations like this:

    child_import_1 status: :failed
      #child_change => parent_import status: :failed
    child_import_2 status: :successful
      #child_change => parent_import status: :successful

In that scenario, even though one of the parent's children had failed,
it still gets a successful status at the end. To prevent that situation,
don't update the parent's status if it's already been set to a finished
status.

Refs #3509, #3511