ParentImportNotifier: Use `Import.finished_statuses`
Instead of manually defining the finished statuses, use the method
that's already defined on `Import` to get them.

The only unfortunate part is that we now had to muss up our `Import`
code and convert those methods to class methods in order for them to be
accessible to `ParentImportNotifier` as well as its instance methods.

Refs #3511