RoutesController: Add separate method for POST /duplicate
Send POST requests for duplication to a new controller method so we can
more easily handle them and separate setup and creation code.

Note that the test doesn't work. The `to change` part succeeds even
though it shouldn't. Still need to figure out what's going on there, but
committing this to hand off the feature to Robert.

Refs #4189