WorkbenchImportWorker#params: Remove unpermitted `referential_id`
This param isn't permitted by the controller and we're currently not
using it for anything. It produces a warning message in the logs, so I
figured we should remove it. We may want to change the architecture to
create referentials outside of the API, and thus would have to add this
param in, but until then it's not needed.