ReferentialMetadataKludge: Get line IDs by `objectid`
Instead of hard-coding an ID for the line IDs, get them by their
`objectid` field.

Worked on this with Alban.

Even though the Kludge class is for testing purposes only and isn't
meant to be real, let's take out the hard-coded values.

The `objectid`s correspond to those in the OFFRE_... zip in
`spec/fixtures/`.