NetexImportsController spec: Fix failing tests caused by Kludge
The `ReferentialMetadataKludge` expects a couple of `Chouette::Line`s to
already exist in the database in order to successfully create a
`ReferentialMetadata` record.

In tests, this is not the case. As a kludge to the kludge, create the
expected line records in the tests that fail because of it, with the
proviso that these additions should be removed or changed once we have
real `ReferentialMetadata` handling for imports.