Import: Remove `#ready?`
This method is no longer used, now that we have Alban's `#update_status`
method.

Originally this was supposed to be used to determine if the import was
finished or not. Now, the parent import status field update doesn't
depend on it, and we can use the `Referential#ready` field to determine
whether or not to display the import in the interface.