Import#update_status: Update `ended_at` when import has finished status
Previously we were only updating the `ended_at` field when the import
had a 'successful' status. But there are other statuses that indicate
the import finished, and the `ended_at` field should be updated in these
cases also.

Committing what I have now which seems to work but I didn't write specs
for it because it was a pain to test (what with creating a child, etc.).
The `#update_status` method should be refactored to make testing the
`ended_at` update easier.

This needs to be committed without specs now because I'm off to go to a
client meeting and this work/branch is blocking QA of imports both
internally and externally. So I guess doing things right comes later.
That is to say, probably a lot later. Who knows.