Import#child_change: Remove `child` argument
It isn't used any more so get rid of it. I had forgotten to remove the
argument from the place it was actually called in the code when I did
this previously in 61817b2e7828455021bcec66e2d6da70f879e1c4. Actually,
Alban hadn't modified the method signature in the code he gave me, I
shouldn't have committed that change as him, my bad.

The argument was added back in because it caused `#notify_parent` to
fail. Update that method so it makes the correct method call.