TableBuilderHelper::Column: Add `link_to` argument
This parameter will be used as the `href` to link the column value
somewhere.

We give it a lambda because this makes it easier to call any method on
the row object.

This means the accessor needs to take the object as an argument, like in
the `#value` method, because we don't have a better way to handle that
(it can't be done at initialisation time because at that point we don't
have row objects, we have a collection).