TableBuilderHelper#tbody: Use new `Column#link_to` for all column links
Instead of always linking to the `name` and `comment` attributes, allow
any column to link to any thing by calling the new
`TableBuilderHelper::Column#link_to` method.

Because `linkable?`-ness is now determined at the column level, we can
get rid of the `#column_is_linkable?` method here.

The code is still pretty much "en désordre" as a result of the
`overhead` code tornado, but we'll have to deal with that later. (Oh
goodness I hope we get to deal with it.)