table_builder_2: Use `Column#link_to` in all table builder calls
Rewrite table builder calls to work with the new
`TableBuilderHelper::Column#link_to` mechanism for defining the linked
column and what it points to. This takes the link out of the
`TableBuilderHelper`, and puts it into the hands of the view.

It gives us more flexibility with regards to what can be defined as a
link in our tables and what those links point to.

This will be used in a more meaningful way in the `imports#show` page to
list `NetexImport`s and provide links to both the referential associated
with a Netex import and the imported file status page from a table row.