Revert "TableBuilderHelper::Column#link_to: Allow arbitrary number of arguments"
This reverts commit 28db706443a912e8355e4c48488dc40c403e7f76.

Turns out we didn't need to be able to pass an arbitrary number of
arguments to the lambda after all. The URL helper objects necessary in
addition to the first argument to the lambda can be retrieved from the
view context directly instead of passing them into the block as
parameters (which would actually make things more difficult, because the
block is called in the `TableBuilderHelper`, thus outside the scope of
the view).