WorkbenchImportWorker spec: Use `subdirs` in place of `entry_groups`
Since the concept of `entry_groups` was changed to `subdirs`, this
change needs to be reflected in the tests.

Instead of looping over `entry_groups`, we now loop over `subdirs`, and
pass this object directly to `mock_post`.

The `mock_post` helper is now rewritten to reflect the new structure
of post parameters. Remove the `params` `let` because it's not used
anywhere else and we need to pass the arguments to `mock_post` into it.

This currently errors because `Faraday::UploadIO` isn't passed the
parameters that are expected in `mock_post`.