WorkbenchImportWorker spec: Check `ended_at` update after import
The import job now updates its `ended_at` time at the end of the import
job. The mock didn't know this, though, so it gave us an error at that
call. Add a new expectation for the update of the `ended_at` time to get
this test to pass again.