ImportMessage: Remove `dependent: :destroy` on :resource association
It doesn't really make sense to have a `dependent: :destroy` here. For
one thing, the `Import` class has the necessary `dependent: :destroy`
definitions to remove its associated objects when it gets deleted.

At this point, according to Luc, we don't envision deleting
`ImportResource`s separately from `Import`s.

The other weird thing is that:

* `ImportResource` has many `ImportMessage`s
* `ImportMessage` has one `ImportResource`

With the `dependent: :destroy` set up the way it is here on
`ImportMessage`, we could potentially be deleting an `ImportResource`
with other `ImportMessage` children that would then become orphaned.

For both these reasons, it seems better to remove this trigger.

Refs #4412