Add `ModelAttribute`
This new class will allow us to generate a list of all editable fields
in all our models including the type of each field.

We need this for Compliance Control, in order to get a list of fields or
models & fields that can be selected to validate a given model
validation check. The crucial bits here are the models, fields, and the
types of those fields. These need to be defined (manually at least to
begin with), accessible, and filterable.

Refs #4401